New Clusterize function
 Committer: Erik Wright <eswright@pitt.edu>
	modified:   DESCRIPTION
	modified:   NAMESPACE
	modified:   R/AlignProfiles.R
	new file:   R/Clusterize.R
	modified:   R/DB2Seqs.R
	modified:   R/DistanceMatrix.R
	deleted:    R/IdClusters.R
	modified:   R/IdLengths.R
	modified:   R/SearchDB.R
	new file:   man/Clusterize.Rd
	modified:   man/DB2Seqs.Rd
	modified:   man/DECIPHER-package.Rd
	modified:   man/DistanceMatrix.Rd
	deleted:    man/IdClusters.Rd
	modified:   man/IdLengths.Rd
	modified:   src/AlignProfiles.c
	modified:   src/DECIPHER.h
	modified:   src/DistanceMatrix.c
	modified:   src/EnumerateSequence.c
	modified:   src/MultiMatch.c
	new file:   src/PairwiseAlignment.c
	modified:   src/R_init_decipher.c
	new file:   src/SphericalKmeans.c
	new file:   vignettes/ClusteringSequences.Rnw
	modified:   vignettes/DECIPHERing.Rnw
	modified:   vignettes/DesignPrimers.Rnw
	modified:   vignettes/DesignSignatures.Rnw
	modified:   vignettes/FindChimeras.Rnw
	modified:   vignettes/FindingGenes.Rnw

+++ src/AlignProfiles.c
@@ -657,14 +657,14 @@ SEXP alignProfiles(SEXP p, SEXP s, SEXP type, SEXP subMatrix, SEXP dbnMatrix, SE
 		if (i % (ls + 1) == 0) {
 			Rprintf("\n");
 		}
-		Rprintf("%d ", (int)m[i]);
+		Rprintf("%1.1f ", m[i]);
 	}
 	Rprintf("\n");
 	for (i = 0; i < lp*ls; i++) {
 		if (i % ls == 0) {
 			Rprintf("\n");
 		}
-		Rprintf("%d ", (int)o[i]);
+		Rprintf("%d ", o[i]);
 	}
 	Rprintf("\n");
 	*/
@@ -1484,14 +1484,14 @@ SEXP alignProfilesAA(SEXP p, SEXP s, SEXP subMatrix, SEXP hecMatrix, SEXP go, SE
 		if (i % (ls + 1) == 0) {
 			Rprintf("\n");
 		}
-		Rprintf("%d ", (int)m[i]);
+		Rprintf("%1.1f ", m[i]);
 	}
 	Rprintf("\n");
 	for (i = 0; i < lp*ls; i++) {
 		if (i % ls == 0) {
 			Rprintf("\n");
 		}
-		Rprintf("%d ", (int)o[i]);
+		Rprintf("%d ", o[i]);
 	}
 	Rprintf("\n");
 	*/
